requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
lxml
playwright
mysql-connector-pythonpython-dotenv